% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc.R
\name{vacc}
\alias{vacc}
\title{Tabla de frecuencias para VACC con gráficos opcionales}
\usage{
vacc(
  x,
  histograma = FALSE,
  ojiva = FALSE,
  hyp = FALSE,
  color_relleno = "#E6E6E6",
  color_linea = "#0099CC",
  color_borde = "#333333"
)
}
\arguments{
\item{x}{Vector numérico.}

\item{histograma}{Sí es TRUE muestra el histograma de frecuencias. Por defecto es FALSE}

\item{ojiva}{Sí es TRUE muestra el gráfico de ojiva. Por defecto es FALSE}

\item{hyp}{Sí es TRUE muestra el histograma de frecuencias con polígono de frecuencias. Por defecto es FALSE}

\item{color_relleno}{Color de relleno}

\item{color_linea}{Color de la línea}

\item{color_borde}{Color de borde}
}
\value{
Genera una lista con 4 elementos:
\enumerate{
  \item Tabla de frecuencias
  \item Histograma de frecuencias
  \item Gráfico de ojiva
  \item Histograma y Polígono de frecuencias
}
}
\description{
Genera la tabla de frecuencias de una variable aleatoria cuantitativa continua (VACC). A su vez, genera gráficos opcionales como histograma, ojiva, y polígono de frecuencias. Por defecto, se define el número de clases con la fórmula de Sturges.
}
\examples{
# Devuelve tabla de frecuencias
vacc(x = iris$Sepal.Length)
# Tabla de frecuencias + histograma de frecuencias
vacc(x = iris$Sepal.Length, histograma = TRUE)
# Tabla de frecuencias + gráfico de ojiva
vacc(x = iris$Sepal.Length, ojiva = TRUE)
# Tabla de frecuencias + histograma y polígono de frecuencias
vacc(x = iris$Sepal.Length, hyp = TRUE)

}
